feat(diet,seat): 수정사항 반영
