+++ src/routers/v1/index.js
@@ -1,12 +1,12 @@
 import express from 'express';
 import facility from './facility';
-import sMeal from './sMeal';
-import libSeat from './libSeat';
+import sMeal from './univDiet';
+import libSeat from './librarySeats';
 
 const router = express.Router();
 
 router.use('/facility', facility);
-router.use('/sMeal', sMeal);
+router.use('/diet', sMeal);
 router.use('/libSeat', libSeat);
 
 export default router;
