--- src/routers/v1/sMeal.js
@@ -1,21 +1,0 @@
-import express from 'express';
-import serGetSMealResponse from '../../services/serGetSMeal';
-import serMakeSMealResponse from '../../services/serMakeSMeal';
-
-const router = express.Router();
-
-router.get('/get', (req, res) => {
-  serGetSMealResponse(req.query.time, req.query.version).then(function (text) {
-    // 성공시
-    let returnToSkill = serMakeSMealResponse(text[0], text[1]);
-
-    res.json(returnToSkill);
-  }, function (error) {
-    // 실패시 
-    console.error(2);
-    res.json(error);
-  });
-
-});
-
-export default router;
